LLPTW, RVH: fix the mem_resp_hit when the new req's status is changed to last_hptw_req (#2983)
When the req is sent into LLPTW and its status need to be changed to
last_hptw_req, the mem_resp_hit is not valid. It make L2TLB don't store
the pte of the req. That is why L2TLB resp a wrong stage 1 pte. So the
mem_resp_hit need to be valid when the new req's status is changed into
last_hptw_req directly.